Create actionAddDec in utils, for refactoring
