+++ front-end/src/utils/actionAddDec.js
@@ -1,0 +1,19 @@
+const decrement = (event) => {
+  const idProduct = event.target.id.split('-')[1];
+  const qtdProduct = document.getElementById(`qtd-${idProduct}`).value - 1;
+  if (qtdProduct >= 0) {
+    document.getElementById(`qtd-${idProduct}`).value = qtdProduct;
+  }
+};
+
+const increment = (event) => {
+  const idProduct = event.target.id.split('-')[1];
+  let qtdProduct = parseInt(
+    document.getElementById(`qtd-${idProduct}`).value,
+    10,
+  );
+  qtdProduct += 1;
+  document.getElementById(`qtd-${idProduct}`).value = qtdProduct;
+};
+
+module.exports = { decrement, increment };
